apps/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
# SQLAlchemy >=2.0.30 soporta Python 3.13 (evita fallo TypingOnly).
sqlalchemy[asyncio]==2.0.37
psycopg[binary]==3.3.2
aiosqlite==0.19.0
alembic==1.13.1
# Pydantic 2.12.5 ships pydantic-core 2.41.x con cp313 wheels (Python 3.13).
pydantic==2.12.5
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-genai==0.5.0
python-multipart==0.0.9
